Addressed all pull request comments
